refactor: Afternote 에디터 관련 상태 클래스 패키지 이동 및 구조 정리
- `AfternoteEditorState`, `MemorialPlaylistStateHolder`, `AfternoteSaveState` 등 상태 관리 클래스들을 `model` 패키지에서 `state` 패키지로 이동
- `AfternoteEditorState` 내 불필요한 전체 경로 임포트를 제거하고 `InformationProcessingMethod`, `RegisterAfternotePayload` 등에 대한 임포트 최적화
- `AfternoteEditorState`에 수신자 목록이 비어있을 때 외부 데이터를 채워넣는 `replaceFromAuthorDirectoryIfEmpty` 함수 추가
- 사용되지 않는 `DataProviderModule.kt` 삭제
- 패키지 구조 변경에 따라 관련 컴포저블 및 ViewModel의 임포트 경로 업데이트
- `ProcessingMethodManager` 주석 내 클래스 참조 경로 수정
