+++ feature/afternote/presentation/src/main/kotlin/com/afternote/feature/afternote/presentation/author/editor/AfternoteEditorViewModel.kt
@@ -15,16 +15,19 @@ import com.afternote.feature.afternote.domain.usecase.author.editor.UpdateUseCas
 import com.afternote.feature.afternote.domain.usecase.author.editor.UploadMemorialPhotoUseCase
 import com.afternote.feature.afternote.domain.usecase.author.editor.UploadMemorialThumbnailUseCase
 import com.afternote.feature.afternote.domain.usecase.author.editor.UploadMemorialVideoUseCase
+import com.afternote.feature.afternote.presentation.author.editor.cache.EditorAuthorReceiversUiCache
+import com.afternote.feature.afternote.presentation.author.editor.mapper.AfternoteEditorMapper
+import com.afternote.feature.afternote.presentation.author.editor.mapper.CreateInput
+import com.afternote.feature.afternote.presentation.author.editor.mapper.MemorialMediaUrls
 import com.afternote.feature.afternote.presentation.author.editor.model.AfternoteEditorReceiver
-import com.afternote.feature.afternote.presentation.author.editor.model.AfternoteEditorState
-import com.afternote.feature.afternote.presentation.author.editor.model.AfternoteSaveState
-import com.afternote.feature.afternote.presentation.author.editor.model.AfternoteValidationError
-import com.afternote.feature.afternote.presentation.author.editor.model.AfternoteValidationException
 import com.afternote.feature.afternote.presentation.author.editor.model.EditorCategory
-import com.afternote.feature.afternote.presentation.author.editor.model.MemorialPlaylistStateHolder
 import com.afternote.feature.afternote.presentation.author.editor.model.RegisterAfternotePayload
 import com.afternote.feature.afternote.presentation.author.editor.playlist.Song
-import com.afternote.feature.afternote.presentation.author.editor.provider.EditorAuthorReceiversUiCache
+import com.afternote.feature.afternote.presentation.author.editor.state.AfternoteEditorState
+import com.afternote.feature.afternote.presentation.author.editor.state.AfternoteSaveState
+import com.afternote.feature.afternote.presentation.author.editor.state.AfternoteValidationError
+import com.afternote.feature.afternote.presentation.author.editor.state.AfternoteValidationException
+import com.afternote.feature.afternote.presentation.author.editor.state.MemorialPlaylistStateHolder
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.flow.MutableStateFlow
