--- feature/afternote/presentation/src/main/kotlin/com/afternote/feature/afternote/presentation/author/editor/provider/DataProviderModule.kt
@@ -1,9 +1,0 @@
-package com.afternote.feature.afternote.presentation.author.editor.provider
-import dagger.Module
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-
-@Module
-@InstallIn(SingletonComponent::class)
-@Suppress("UNUSED") // Used by Hilt via @InstallIn
-object DataProviderModule
