+++ feature/afternote/presentation/src/main/kotlin/com/afternote/feature/afternote/presentation/author/navigation/AfternoteNavGraphAddSong.kt
@@ -2,10 +2,10 @@ package com.afternote.feature.afternote.presentation.author.navigation
 
 import androidx.compose.runtime.Composable
 import androidx.navigation.NavController
-import com.afternote.feature.afternote.presentation.author.editor.model.MemorialPlaylistStateHolder
 import com.afternote.feature.afternote.presentation.author.editor.playlist.AddSongCallbacks
 import com.afternote.feature.afternote.presentation.author.editor.playlist.AddSongScreen
 import com.afternote.feature.afternote.presentation.author.editor.playlist.AddSongViewModel
+import com.afternote.feature.afternote.presentation.author.editor.state.MemorialPlaylistStateHolder
 
 @Composable
 internal fun AfternoteAddSongNavigation(
